code/embedding-search/ingestion/cloud-run/ingestion-trigger-handler-service/requirements.txt
Flask==3.0.0
google-cloud-run==0.10.0
cloudevents==1.10.1
gunicorn==21.2.0